small fix (#88)
